email-assistant/backend/requirements.txt
# Core
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary

# Email
imaplib2==3.6

# Scheduler
APScheduler==3.10.4

# ML
scikit-learn==1.3.2
nltk==3.8.1
beautifulsoup4==4.12.2
joblib==1.3.2
scikit-learn==1.3.2

# AI Models
transformers==4.38.0
torch==2.2.0
datasets==2.18.0
sentencepiece==0.1.99
accelerate==0.27.0

# Utils
numpy==1.26.4
pandas==2.1.4